Action query working

+++ server/schemas/typeDefs.js
@@ -7,9 +7,9 @@ const typeDefs = gql`
   type Impact {
     _id: ID
     date: Date
-    travelContribution: Int
-    energyContribution: Int
-    foodWasteContribution: Int
+    travelContribution: Float
+    energyContribution: Float
+    foodWasteContribution: Float
   }
 
   type User {
@@ -17,7 +17,7 @@ const typeDefs = gql`
     username: String
     email: String
     password: String
-    impactScore: Int
+    impactScore: Float
     dailyImpact: [Impact]
   }
 
@@ -25,7 +25,7 @@ const typeDefs = gql`
     _id: ID
     name: String
     units: String
-    carbonPerUnit: Int
+    carbonPerUnit: Float
     category: String
   }
